requirements.txt
agate==1.9.1
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.4.0
babel==2.17.0
backports.tarfile==1.2.0
boto3==1.42.17
botocore==1.42.17
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
daff==1.4.2
dbt-adapters==1.22.2
dbt-common==1.37.2
dbt-core==1.11.2
dbt-extractor==0.6.0
dbt-protos==1.0.410
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.11.0
deepdiff==8.6.1
filelock==3.20.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.7.0
leather==0.4.1
logging==0.4.9.6
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.6
numpy==2.4.0
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.5.1
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==4.1.1
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
sqlparse==0.5.4
text-unidecode==1.3
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
zipp==3.23.0
